requirements.txt
python-dotenv==0.14.0
dynaconf==3.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
SQLAlchemy-serializer==1.3.4.2
SQLAlchemy-ImageAttach==1.1.0
psycopg2==2.8.6
Flask-Cors==3.0.9
bcrypt==3.2.0
PyJWT==1.4.2
qrcode[pil]==6.1
eventlet==0.29.1
Flask-SocketIO==4.3.1
